requirements.txt
apache-airflow
psycopg2-binary
kafka-python